Create sudoku board filling mechanisms
Player can now fill in sudoku board with numbers. The working
functions are still experimental, design is terrible aswell, so those
things have to be improved definitely. At that time, player is able to
try completing sudoku, but it has not been set strict rules of how to
place initial numbers on the board.