Create sudoku backtracking algorithm
Backtracking has been finally created, but it needs some modifications.
It is capable of showing proper sudoku solution, but just one. In near
future it has to gether all possible solutions to the board, and
giving back a note whether sudoku has one or more solutions. Long run
it will help detecting and describing, what conditions are that makes
sudoku with multiple solutions, and thus better dealing with and
avoiding them.