Improve engine to create only unique sudokus
The engine is capable of creating sudokus, that are unique every
single render. However, the remarkable tradeoff is that the
difficulties have been nullified by above mentioned change (it is not
really possible to choose the initial digit count and force sudoku to
keep those while being unique). Thankfully, there is an actual option,
which forces computer to solve sudoku using different known techinques,
and giving 'scores' per difficulty and it makes them more maintainable.