Apply Single Position mechanism
From now on, Sudoku engine is not only capable of filling gaps that
match Single Candidate rules, but it works for Single Position aswell.
Engine seems to be working well for now - but it is still necessary to
apply some harder solving concepts to the enginge, and make it iterable
(going from easiest to hardest method - if a method helps finding
a digit - by applying digit or removing options - go back to the easiest
method; if did not help -> then find a harder method. If no method
found, return some note...)