Prepare Naked Subset First General Rule Part (err)
Naked Subset Fundamental rules has been applied to the engine.
The main rule states, that whenever we have x tiles with x options in it
and that options are all the same for those tiles, that means other
tiles would not have those digits, and they can be removed from them.
Mechanism is applied for double and triple subsets - quad subsets are
incredibly rare, so it is not obligatory to implement quads (for now at
least). However, there's some double and tripple specific parts that has
not been added yet + even this main rule seems to throws error, so
getting rid of it is the top priority now.