Develop part two for naked subset
Naked subset part two functions are now prepared for deployment.
Currently working on "incomplete tripple subset" one, which is going
quite well and simple so far. Apart from that, in order to finish naked
subset method, it is also necessary to create "tripple double subset"
function and append it to naked subset method.