Finish incomplete tripple subset, naked-S function
The second part of naked Subset method has been implemented and tested
and it works great. No issues, simply does what it's supposed to. to
finish this method off, the very last part - tripple double subset -
needs to be implemented. It's the last helper function for naked Subset
method. After completing it, the hidden Subset would be the main deal.