Finish naked subset method
Naked subset is finally completed ! All of three method parts do the
right job without throwing errors. This method was incredibly time
consuming, since it was like 3 little methods in one. After tough
issues and errors, this can be sign as completed. Now it's time to
go for the last method marked as hard, which is Hidden Subset.