Complete hidden Subset function-Incomplete tripple
After lots of errors, the very first part of hidden Subset is ready to
go. Incomplete tripple subset is finally implemented. It's a very
case-specific method, which works rarely, but seems to be very helpful.
For that method I've implemented new testing method, which allows to
pass any grid to it, and check if for that specific grid currently
developed function helps. It might be very handy for more complex
solving methods. Now last two parts for hidden Subset is: 1) to focus
on it's main ability (checking for three tripples or two doubles of any
length, but with three or two unique numbers), and to implement
three double subset (same here - we don't care about its' length,
but we care if three subsets have three unique digits, that are
present twice for dimension, and only in two of three potential subset
elems).