Finish hidden Subset double/tripple normal part
The basic part of hidden Subset has been implemented. It has been
tested very strictly to ensure it works properly. Hidden Subset in total
is almost done - the very last function to provide is actually
Three Double Subset - in case of hidden version, we don't care about
the length of those three subsets, but we check if three of them
contains three unique digits (two of three for every element), and if
we found something, we update only the subset elems.