Finish Hidden Subset Method
Hidden Subset (and all his sub-functions) has been finally completed.
Three double hidden subset seems not to be testable as hidden normal
subset can do part of its' job. The conclusion might be that TDHS
either would never be useful or only useful for very rare, specific
cases. It doesn't seem to be throwing errors, so it is fine for now.
From now on hard-rated methods are completed, and now we can move on
to implementing X-wings, which is first method rated as extreme
difficulty.