Prepare X-wings / Swordfish initial variables
Main preparation for x-wings and swordifs methods are done! We have
two properly working variables. First indicates which numbers in
current lines exist exactly twice as an option; second one indicates
the number indexes (in which tiles in line this number is as an option).
It also means that those variables will always have the same length.
Important nore is that Hidden Three Doubles has been detected to throw
errors! For now it has been nullified and this very last part of hidden
subset is now longer trying to remove digits. This hidden subset part
requires some deeper insight into potential issue that lead to this
error. But it's not very important (as long as it is nullified).