Finish X-wings
X-wings are now created. It has been tested using 2 specific cases
(one for row and one for column) and they are not throwing any
errors - they are finding the right number apart ! That's pretty
sure it's written 100% correctly. At last it is necessary to
finish off with Swordfish method, which would be the last method
to implement (at least for now). It is very similiar to X-wings, but
a little more difficult conceptually and coding-wise.