Complete SwordFish method
SwordFish has been completed ! After a few tests done, it's also
working properly! There's no known errors for now, and I will take a
closer look if it remains as it is for now. Now on, all necessary
methods are completed ! The ones ommitted are:
1) Three doubles (a hidden subset sub-function) -> throws errors, so it
was commented out for now;
2) Forcing chains - maybe would be implemented in the future;
Next part to finish huge engine part, is to make some time optimisations
in order to make sudoku grids and proper difficulty adjustment real
quick.