Engine approach modified
This part was abslutely tremendous ! Involved a lot of issues around
time execution when trying to set sudoku difficulty to what player did
choose. Going forward was making things much worse, so I've taken a step
backwards. Time execution is now much more stable (sudoku usually
renders under 1 second), but there is a tradeoff for that: a sudoku
with random difficulty is being rendered. For now it means, player can
not choose Sudoku difficulty, however Sudoku is able to determine the
difficulty of rendered board, so player would get notified about that.
In the future I'll try to apply this difficulty choose to the game (with
actually proper time rendering). Currently there would be some extra
adjustments to do regarding this functionality change.