Completed pencilmarks
Pencilmarks is the very first feature introduced as a player solving
tools. It enables to mark a tile for possible candidates (which might
be espacially useful for harder Sudokus). Marks can be removed only
when pencilmarks mode is off. The basic mechanism is fully completed,
however there are two main aspects, that can be improved there:

1) When applying number to a tile, remove it's pencilmarks from the same
row, column and square;

2) Improve animations when adding new pencilmark - adding new pencilmark
to a tile, which already have some pencilmarks, should force the
animation to work only for this new pencilmark, not for remain ones.