Resolve history travel related bug
Bug happened when using rubber and then moving through history. It was
possible to append NaN instead of ''. Furthermore, this issue lead to
an error, which crashed the app (most probably that was the reason
for that). However, some insights would be taken to make sure that this
error would never happen again.