Minor refactoring & bug solving
New changes to rendering mechanism has been applied. It no longer relies
on board state, which used to be impractical and could negatively impact
further app improvements. Also it helped solving issue, when player
qucikly resets the board, which as a result was showing completely
solved sudoku for a brief duration.