Add loading screen during very first Sudoku render
When player clicks "Play" button on landing page, it also shows the
loading screen there (as it was only for reset Sudoku button before).
Furthermore, a weird flash during first render has been also neutralized
It was showing part of the board layout, which was suddenly
interrupted by loading screen, which caused the flash.