Solve  history travel bugs
Last changes modifies history travel behaviour a little bit, which
ended up with new bug, which causes history board to render 3 to 4
times at the start, forcing problems while travelling. It is no longer
an issue. Another seen bug was from the past - but it was considered
as solved - so now it probably has been ultimately neutralized. It was
associated with rubber usage and history travel, causes one of history
step tile to have value of null. Instead, when it happens, this value
is converted to NaN, which is acceptable solution, which also indicates
that this tile it's not just a blank tile, but blank tile which has been
modified by a player.