[1.1.0] Add number box digit usage limit
This commit contains following changes:

1) When each digit is used exactly 9 times in a Sudoku board, it can no
longer be used. It's a quality change, since there is no option that
Sudoku has more than 9 places for a digit. When limit is reached,
player cannot use this number in number box. He can undo this by
history travel or using rubber. Pencilmarks do not count to the limit.