Complete digit conflicts functionality
Digit conflicts mechanism received a final update. Now our board is able
to display conflict digits by a slight background gradient apply for all
conflict tiles. This feature is also fully compatible with history
travel feature. The main idea of conflict mechanism is to warn player
whenever his action (adding a digit) is contradictory with Sudoku rules.
This would be an in-built feature, meaning that it will not appear as a
Landing page option, so that it cannot be disabled by a player.

Also some minor design modifications are made for the rubber, Sudoku
board borders, and a pencilmark toolbox option.