Update Sudoku page and dialog boxes
Sudoku page got some new design changes to give it a refreshed, more
fancy styling. Here are all the updates:

- Both Sudoku title and Numbebox digit icons received a slight shadow,
which color is based on a Sudoku difficulty;

- Fixed the background colors for both light and dark modes. The main
goal was to lower their intensity and to cast a light, contrast shadow
around the screen borders;

- Affected Sudoku reset button by rounding it left and right borders,
and hovering over this button light it up a bit;

- Timer (when enabled) has received a bigger font size;

Updates are made for Dialog boxes aswell:

- All the dialog boxes got their borders less rounded (to just 5%), to
revert them back to their original, rectangle-based shape.

- For Info dialog box, a change was made to a close button - once it is
hovered, it fires a slight animation that scales the text by a minor
amount and also applies an underline to it. Moreover, the background
for dark and light themes got changed to be less intensive.

- For Reset dialog box, a fancy hover effect is added to both action
buttons (lighting and scaleing them up, when the cursor is on them);

- For Wind (completion) box, applied a light & scale effect (very
similiar to what is changed for Reset box), and furthermore lower
the background intensity - applied to all difficulty colors.

And some bugs are now solved:

- Fixed the issue, when hovering over currently inactive history travel
button caused cursor to change it look to pointer. This caused a
misleading perspective, that an inactive history button is still
clickable;

- When player wanted to replace a Sudoku with a new one (while keeping
some conflicted tiles), the visual mark was still seen on affected tiles
in new board. That mentioned issue got resolved.